bgp-tool/requirements.txt
Flask
netmiko
python-dotenv
requests
pymongo